# New: no-double-assignments (fixes #3481)