# Performance characteristics for this implementation

Before:

```
CPU Speed is 1800 with multiplier 7500000
Performance Run #1:  3165.737594ms
Performance Run #2:  3165.387589ms
Performance Run #3:  3156.572153ms
Performance Run #4:  3146.225769ms
Performance Run #5:  3129.534463ms
Performance budget ok:  3156.572153ms (limit: 4166.666666666667ms)
```

After:

```
CPU Speed is 1800 with multiplier 7500000
Performance Run #1:  3396.5271119999998ms
Performance Run #2:  3350.931596ms
Performance Run #3:  3398.065597ms
Performance Run #4:  3307.97695ms
Performance Run #5:  3289.81201ms
Performance budget ok:  3350.931596ms (limit: 4166.666666666667ms)
```

Perhaps the code could be improved further, as the hit is noticable. 

It seems a lot of code in this PR (`getPrevNode`, `expressionname`) could be usefully moved into a utility belt on the context or ast-utils. For an example of such a proposal, see #3460.