@denis-sokolov 2015-08-27
Thank you, @nzakas, for your comments.
I did not perfect the code until I had no idea if there's any interest in the rule whatsoever.
I will clean it up, and maybe change the algorithm based on inputs from #3481 and then post an update here.